third_party/WebKit/LayoutTests/fast/dom/shadow/content-reprojection-fallback-crash.html

<!DOCTYPE html>

<html>
<script src="../../../resources/js-test.js"></script>

<p>A crash should not happen when fallback elements are reprojected.</p>

<div id="host1"></div>

<pre id="console"></pre>

<script>
var shadowRoot1 = host1.createShadowRoot();
shadowRoot1.innerHTML = '<div id="host2"><content select="span"><div id="fallback">fallback element</div></content><div id="div">This should be selected.</div></div>';

var host2 = shadowRoot1.getElementById('host2');
var shadowRoot2 = host2.createShadowRoot();
shadowRoot2.innerHTML = '<content select="div"></content>';

var content = shadowRoot2.querySelector('content');
var distributedNodes = content.getDistributedNodes();

shouldBe('distributedNodes.item(0)', 'shadowRoot1.getElementById("fallback")');
shouldBe('distributedNodes.item(1)', 'shadowRoot1.getElementById("div")');

finishJSTest();
</script>

</html>
